// packages/fe-api-client/src/schema.gen.ts
export interface operations {
// ...
    "words.counts": {
        parameters: {
            query?: {
                q?: string;
                status?: "pending" | "running" | "succeeded" | "failed";
            };
            header?: never;
            path: {
                language: "ru" | "en" | "es" | "fr" | "de" | "zh" | "ja" | "hi" | "ar" | "uk";
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        total: number & unknown;
                        pending: number & unknown;
                        running: number & unknown;
                        succeeded: number & unknown;
                        failed: number & unknown;
                    };
                };
            };
        };
    };
}